requirements.txt
blinker==1.9.0
branca==0.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cssselect2==0.8.0
cycler==0.12.1
Flask==3.1.0
flask-login==0.6.3
flask-bcrypt==1.0.1
folium==0.19.4
fonttools==4.55.6
geopandas==1.0.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pdfrw==0.4
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyogrio==0.10.0
pyparsing==3.2.1
pyphen==0.17.2
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.3.1
requests==2.32.3
shapely==2.0.6
six==1.17.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.1
urllib3==2.3.0
weasyprint==65.0
webencodings==0.5.1
Werkzeug==3.1.3
xyzservices==2025.1.0
zopfli==0.2.3.post1
